pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rp-jump"
version = "1.5.0"
description = "Chrome extension server for jumping from ReportPortal to Magna logs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
rpjump-server = "rpjump_server:main"

